DSK/requirements.txt
aiogram==2.25.1
tensorflow==2.12.0
PyYAML==6.0